Reinstate refactoring of wgTurnOn functions, and move building blocks to tun_setup.go
